requirements.txt

biopython>=1.79
pytest>=7.0

